refactor(entities): update table name and column type for clarity and consistency
- Changed table name for `Option` entity to `question_options` for better clarity
- Updated `order` column type in `Question` entity from `int` to `numeric` for consistency
- Simplified sorting logic in `QuestionService` for improved readability

+++ src/survey/entities/question.entity.ts
@@ -18,9 +18,10 @@ export class Question {
 		nullable: false,
 	})
 	type: QuestionType;
-	
-	@Column({ type: 'int', default: 0 })
+
+	@Column({ type: 'numeric', default: 0 })
 	order: number;
+
 	@CreateDateColumn({ type: 'timestamptz' })
 	created_at: Date;
 	@UpdateDateColumn({ type: 'timestamptz' })
